fix(filters): add useMenuList, wrap function in useCallback
