+++ src/modules/teacher/components/AssignmentListTeacherFilters/AssignmentListTeacherFilters.tsx
@@ -1,9 +1,10 @@
 import { FilterOutlined } from '@ant-design/icons';
-import { Dropdown, Menu, Space } from 'antd';
+import { Dropdown, Space } from 'antd';
 import type { MenuProps } from 'antd';
-import { useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { AssignmentsData, Difficulty, FILTER_METHODS } from '../../models';
 import s from './AssignmentListTeacherFilters.module.css';
+import { useMenuList } from '../../../../helpers/useMenuList';
 
 type Props = {
   assignmentsData: AssignmentsData[];
@@ -16,8 +17,6 @@ export const AssignmentListTeacherFilters = ({
   setFilteredData,
   openAssignmentsIds,
 }: Props) => {
-  const [openKeys, setOpenKeys] = useState<string[]>([]);
-
   const filterDifficulty = (difficulty: Difficulty) => {
     setFilteredData(
       assignmentsData.filter((item) => item.difficulty === difficulty),
@@ -73,18 +72,7 @@ export const AssignmentListTeacherFilters = ({
     }
   };
 
-  const onOpenChange: MenuProps['onOpenChange'] = (keys) => {
-    const latestOpenKey = keys.find((key) => openKeys.indexOf(key) === -1);
-    setOpenKeys(latestOpenKey ? [latestOpenKey] : []);
-  };
-  const menu = (
-    <Menu
-      openKeys={openKeys}
-      onOpenChange={onOpenChange}
-      onClick={handleMenuFilters}
-      items={filterItems}
-    />
-  );
+  const menu = useMenuList(filterItems, handleMenuFilters);
 
   return (
     <div className={s.filter}>
