+++ src/modules/teacherLessons/components/TeacherLessonsFilters.tsx
@@ -1,13 +1,11 @@
 import { FilterOutlined } from '@ant-design/icons';
-import { Dropdown, Menu, Space } from 'antd';
-import type { MenuProps } from 'antd';
-import React, { useMemo, useState } from 'react';
+import { Dropdown, Space } from 'antd';
+import React, { useMemo } from 'react';
 import { FILTER_METHODS } from '../models';
 import s from './TeacherLessonsFilters.module.css';
+import { useMenuList } from '../../../helpers/useMenuList';
 
 export const TeacherLessonsFilters = () => {
-  const [openKeys, setOpenKeys] = useState<string[]>([]);
-
   const filterItems = useMemo(
     () =>
       FILTER_METHODS.map((item) => ({
@@ -23,13 +21,7 @@ export const TeacherLessonsFilters = () => {
     [],
   );
 
-  const onOpenChange: MenuProps['onOpenChange'] = (keys) => {
-    const latestOpenKey = keys.find((key) => openKeys.indexOf(key) === -1);
-    setOpenKeys(latestOpenKey ? [latestOpenKey] : []);
-  };
-  const menu = (
-    <Menu openKeys={openKeys} onOpenChange={onOpenChange} items={filterItems} />
-  );
+  const menu = useMenuList(filterItems);
 
   return (
     <div className={s.filter}>
